Add Matrix client for output notifications
This commit introduces a Matrix client to send the scrape results to a Matrix server, laying the groundwork for scheduled notifications.

For security, the client will only accept room invite requests from users on the same homeserver as the configured user.

If Matrix configuration is not provided, the program will continue to print out to the console.

Additional improvements:
- Includes product links in the output for better usability
- Refactors product-related code into a separate file for better separation of concern